backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-dotenv==1.1.0
httpx==0.28.1
pytest==8.3.5
pytest-asyncio==0.25.3
ruff==0.11.6
pyjwt==2.10.1
resend==2.19.0
supabase==2.24.0
websockets>=13.1
pydantic[email]==2.12.5
pydantic-settings==2.11.0
